chore(@clayui/data-provider): update DataProvider component stories
